Tests/LibWeb/Text/expected/scroll-during-page-load.txt

Scroll position: 200px
